refactor: Moving logic from controllers to services
