+++ Modules/Article/zArticleDelete/Config/config.php
