+++ Modules/Article/zArticleDelete/Providers/ArticleServiceProvider.php
