+++ Modules/Article/zArticleDelete/Resources/views/index.blade.php
