+++ Modules/Article/zArticleDelete/Resources/views/layouts/master.blade.php
