--- Modules/Article/zdatabase/seeders/ArticleDatabaseSeeder.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Modules\Article\Database\Seeders;
-
-use Illuminate\Database\Seeder;
-
-class ArticleDatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     */
-    public function run(): void
-    {
-        // $this->call([]);
-    }
-}
